Rename templates to layouts
